Decrease hearing dispersion a bit
This should be about low enough that bots don't look almost directly to gunshots.

SAIN has hearing dispersion tied to a point which a bot uses to investigate a sound source. This has the benefit of a simple, 2-in-1 solution, but also has a drawback of being inflexible, in my opinion.

You'd want large dispersion so bots will look properly confused and not look like they snap to the player's position, However, too much dispersion and you run the risk of fail pathing, which SAIN will fall back to last known location or the player's current position, in that order.